refactor(cognitive_memory/mod): extract tests to sibling (#1266)
Partial: mod.rs 1269 -> 1015 LOC. Source split deferred.
